feat: add games in progress count to home page
